debugged remote code
Not quite sure what was going on or even how we fixed it. There were supposed indentation problems that don't exist. Plus the ir-sensor-code-if-statements were the only things that we added and they seemed to work but the math commands we sent which we never changed did not work...strange?! (ツ)